Fix: rename Navigator Component
